requirements.txt
fastmcp
playwright>=1.40.0
uvicorn[standard]
pydantic
python-multipart